+++ voting-system/src/App.jsx
@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
-import logo from './logo.svg';
+import { BrowserRouter as Router, Route } from "react-router-dom";
 import './App.css';
-import { firebase } from './Firebase.jsx';
+import Login from './Views/Login.jsx';
+import Home from './Views/Home.jsx';
 
 class App extends Component {
   _isMounted = false;
@@ -10,31 +11,10 @@ class App extends Component {
     this.state = {
       data: [],
     };
-    this.testsRef = firebase.firestore().collection('tests');
-  }
-
-  onCollectionUpdate = (querySnapshot) => {
-    const data = [];
-    querySnapshot.forEach((doc) => {
-      var { 
-        apple,
-        name,
-      } = doc.data();
-      data.push({
-          key: doc.id,
-          doc, // DocumentSnapshot
-          apple,
-          name,
-      });
-    });
-    if (this._isMounted) {
-      this.setState({ data });
-    }
   }
 
   componentDidMount() {
     this._isMounted = true;
-    this.unsubscribe = this.testsRef.onSnapshot(this.onCollectionUpdate);
   }
 
   componentWillUnmount() {
@@ -42,31 +22,12 @@ class App extends Component {
   }
 
   render() {
-    var firebasedata = [];
-    for (var i=0; i<this.state.data.length; i++) {
-      firebasedata.push(
-        <p key={i}>
-          {this.state.data[i].apple} {this.state.data[i].name}
-        </p>
-      );
-    }
     return (
       <div className="App">
-        <header className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-          <p>
-            Edit <code>src/App.js</code> and save to reload.
-          </p>
-          {firebasedata}
-          <a
-            className="App-link"
-            href="https://reactjs.org"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Learn React
-          </a>
-        </header>
+        <Router>
+            <Route exact path="/" component={Home} />
+            <Route path="/login" component={Login} />
+        </Router>   
       </div>
     );
   }
